+++ crypto/mem_dbg.c
@@ -451,7 +451,7 @@ int CRYPTO_remove_all_info(void)
 }
 
 static unsigned long break_order_num = 0;
-void CRYPTO_dbg_malloc(void *addr, int num, const char *file, int line,
+void CRYPTO_dbg_malloc(void *addr, size_t num, const char *file, int line,
                        int before_p)
 {
     MEM *m, *mm;
@@ -555,7 +555,7 @@ void CRYPTO_dbg_free(void *addr, int before_p)
     }
 }
 
-void CRYPTO_dbg_realloc(void *addr1, void *addr2, int num,
+void CRYPTO_dbg_realloc(void *addr1, void *addr2, size_t num,
                         const char *file, int line, int before_p)
 {
     MEM m, *mp;
